Only run compression on files that exist
as NSZ soft-crashes sometimes

+++ library/compression.go
@@ -23,51 +23,51 @@ func (lib *Library) compressionWorker() {
 			return
 		case request := <-lib.fileCompressionRequests:
 			//For each requested file, run it through NSZ and check output
-			log.Info().Str("path", request).Msg("Starting compression")
-			err := lib.NSZCompressFile(request)
-			if err != nil {
-				log.Err(err).Msg("NSZ compression failed")
-				//Cleanup output if it made one
-				if len(request) > 3 {
-					newpath := request[0:len(request)-1] + "z"
-					if utilities.Exists(newpath) {
-						if err := os.Remove(newpath); err != nil {
-							log.Error().Err(err).Msg("cleanup output from failing NSZ failed")
+			if utilities.Exists(request) {
+				log.Info().Str("path", request).Msg("Starting compression")
+				err := lib.NSZCompressFile(request)
+				if err != nil {
+					log.Err(err).Msg("NSZ compression failed")
+					//Cleanup output if it made one
+					if len(request) > 3 {
+						newpath := request[0:len(request)-1] + "z"
+						if utilities.Exists(newpath) {
+							if err := os.Remove(newpath); err != nil {
+								log.Error().Err(err).Msg("cleanup output from failing NSZ failed")
+							}
 						}
 					}
-				}
-			} else {
-				log.Info().Str("path", request).Msg("Compression complete")
-				// Check if the source file has been deleted
-				// Check if the expected output file is made
+				} else {
+					log.Info().Str("path", request).Msg("Compression complete")
+					// Check if the source file has been deleted
+					// Check if the expected output file is made
 
-				if !utilities.Exists(request) {
-					//Source file has been deleted, notify library
-					event := &scanRequest{
-						path:             request,
-						isEndOfStartScan: false,
-						isNotifierBased:  true,
-						fileRemoved:      true,
+					if !utilities.Exists(request) {
+						//Source file has been deleted, notify library
+						event := &scanRequest{
+							path:             request,
+							isEndOfStartScan: false,
+							isNotifierBased:  true,
+							fileRemoved:      true,
+						}
+						lib.fileScanRequests <- event
 					}
-					lib.fileScanRequests <- event
-				}
-				//Figure out the output file path
-				newpath := request[0:len(request)-1] + "z"
-				if utilities.Exists(newpath) {
-					//Source file has been deleted, notify library
-					event := &scanRequest{
-						path:             newpath,
-						isEndOfStartScan: false,
-						isNotifierBased:  true,
-						fileRemoved:      false,
+					//Figure out the output file path
+					newpath := request[0:len(request)-1] + "z"
+					if utilities.Exists(newpath) {
+						//Source file has been deleted, notify library
+						event := &scanRequest{
+							path:             newpath,
+							isEndOfStartScan: false,
+							isNotifierBased:  true,
+							fileRemoved:      false,
+						}
+						lib.fileScanRequests <- event
 					}
-					lib.fileScanRequests <- event
 				}
-
 			}
 		}
 	}
-
 }
 
 func (lib *Library) NSZCompressFile(path string) error {
